third_party/blink/web_tests/external/wpt/css/css-backgrounds/animations/border-width-interpolation-expected.txt

This is a testharness.js-based test.
Found 20 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions: property <border-left-width> from [inherit] to [20px] at (-0.3) should be [33px]
  assert_equals: expected "33px " but got "0px "
[FAIL] CSS Transitions: property <border-left-width> from [inherit] to [20px] at (0) should be [30px]
  assert_equals: expected "30px " but got "0px "
[FAIL] CSS Transitions: property <border-left-width> from [inherit] to [20px] at (0.3) should be [27px]
  assert_equals: expected "27px " but got "6px "
[FAIL] CSS Transitions: property <border-left-width> from [inherit] to [20px] at (0.6) should be [24px]
  assert_equals: expected "24px " but got "12px "
[FAIL] CSS Transitions: property <border-left-width> from [inherit] to [20px] at (1.5) should be [15px]
  assert_equals: expected "15px " but got "30px "
[FAIL] CSS Transitions with transition: all: property <border-left-width> from [inherit] to [20px] at (-0.3) should be [33px]
  assert_equals: expected "33px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <border-left-width> from [inherit] to [20px] at (0) should be [30px]
  assert_equals: expected "30px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <border-left-width> from [inherit] to [20px] at (0.3) should be [27px]
  assert_equals: expected "27px " but got "6px "
[FAIL] CSS Transitions with transition: all: property <border-left-width> from [inherit] to [20px] at (0.6) should be [24px]
  assert_equals: expected "24px " but got "12px "
[FAIL] CSS Transitions with transition: all: property <border-left-width> from [inherit] to [20px] at (1.5) should be [15px]
  assert_equals: expected "15px " but got "30px "
[FAIL] CSS Animations: property <border-left-width> from [inherit] to [20px] at (-0.3) should be [33px]
  assert_equals: expected "33px " but got "0px "
[FAIL] CSS Animations: property <border-left-width> from [inherit] to [20px] at (0) should be [30px]
  assert_equals: expected "30px " but got "0px "
[FAIL] CSS Animations: property <border-left-width> from [inherit] to [20px] at (0.3) should be [27px]
  assert_equals: expected "27px " but got "6px "
[FAIL] CSS Animations: property <border-left-width> from [inherit] to [20px] at (0.6) should be [24px]
  assert_equals: expected "24px " but got "12px "
[FAIL] CSS Animations: property <border-left-width> from [inherit] to [20px] at (1.5) should be [15px]
  assert_equals: expected "15px " but got "30px "
[FAIL] Web Animations: property <border-left-width> from [inherit] to [20px] at (-0.3) should be [33px]
  assert_equals: expected "33px " but got "0px "
[FAIL] Web Animations: property <border-left-width> from [inherit] to [20px] at (0) should be [30px]
  assert_equals: expected "30px " but got "0px "
[FAIL] Web Animations: property <border-left-width> from [inherit] to [20px] at (0.3) should be [27px]
  assert_equals: expected "27px " but got "6px "
[FAIL] Web Animations: property <border-left-width> from [inherit] to [20px] at (0.6) should be [24px]
  assert_equals: expected "24px " but got "12px "
[FAIL] Web Animations: property <border-left-width> from [inherit] to [20px] at (1.5) should be [15px]
  assert_equals: expected "15px " but got "30px "
Harness: the test ran to completion.
